# appointments-impl/rest-openapi.yaml
components:
  schemas:
    Appointment:
      description: Represents an appointment.
      properties:
        date:
          format: date-time
          type: string
        id:
          format: int64
          type: integer
        title:
          type: string
      required:
        - title
      type: object
    Owner:
      properties:
        name:
          type: string
info:
  description: ""
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  title: "Appointments"
  version: v1.0
openapi: 3.0.1
paths:
  "/sites/{siteId}/appointments":
    get:
      operationId: getSiteAppointmentsPage
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            format: int64
            type: integer
        - in: query
          name: filter
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
      responses:
        200:
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Appointment"
                type: array
            application/xml:
              schema:
                items:
                  $ref: "#/components/schemas/Appointment"
                type: array
          description: "Get appointments"
      tags: ["Appointment"]
    post:
      description: ""
      operationId: postSiteAppointment
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            format: int64
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Appointment"
          application/xml:
            schema:
              $ref: "#/components/schemas/Appointment"
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
            application/xml:
              schema:
                $ref: "#/components/schemas/Appointment"
          description: ""
      tags: ["Appointment"]
  "/appointments/{appointmentId}":
    get:
      operationId: getAppointment
      parameters:
        - in: path
          name: appointmentId
          required: true
          schema:
            format: int64
            type: integer
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
            application/xml:
              schema:
                $ref: "#/components/schemas/Appointment"
          description: "Get appointment"
      tags: ["Appointment"]
    delete:
      description: Deletes an appointment
      operationId: deleteAppointment
      parameters:
        - in: path
          name: appointmentId
          required: true
          schema:
            format: int64
            type: integer
      responses:
        204:
          content:
            application/json: {}
      tags: ["Appointment"]
    put:
      description: Updates an appointment
      operationId: putAppointment
      parameters:
        - in: path
          name: appointmentId
          required: true
          schema:
            format: int64
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Appointment"
          application/xml:
            schema:
              $ref: "#/components/schemas/Appointment"
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
            application/xml:
              schema:
                $ref: "#/components/schemas/Appointment"
          description: ""
      tags: ["Appointment"]
  "/appointments/{appointmentId}/owner":
    get:
      operationId: getAppointmentOwner
      parameters:
        - in: path
          name: appointmentId
          required: true
          schema:
            format: int64
            type: integer
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
            application/xml:
              schema:
                $ref: "#/components/schemas/Owner"
          description: ""
      tags: ["Appointment"]